refactor(plugins): migrate all panels, stories, and tests into plugin dirs
- Inline remaining panel components (moodTones, socialSharing, map, arrivalCanvas,
  treevites, valenceBeatPad) directly into their plugin dirs; delete panels/ originals
- Move all stories from stories/ into co-located plugin dirs; fix relative imports
- Move all plugin server tests from tests/ into co-located plugin dirs; fix imports
- Move NeighborPanel component test and imageCanvas coords test into plugin dirs
- Update vitest config to include plugins/**/*.test.tsx in component test project
- Fix broken import paths in treevites and valenceBeatPad components
