+++ plugins/light/LightShow.stories.tsx
