+++ plugins/signatureCanvas/component.stories.tsx
