+++ plugins/socialSharing/component.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 import { expect, within } from 'storybook/test';
 import React from 'react';
-import SocialMediaPanel from '../app/components/panels/SocialMediaPanel';
-import { SocialMediaConfigProvider } from '../app/context/PanelConfigs';
-import type { SocialConfig } from '../app/types';
+import SocialMediaPanel from './component';
+import { SocialMediaConfigProvider } from '../../app/context/PanelConfigs';
+import type { SocialConfig } from '../../app/types';
 
 const EMPTY_CONFIG: SocialConfig = { default: '', twitter: '', bluesky: '', mastodon: '', instagram: '' };
 const DEFAULT_ONLY_CONFIG: SocialConfig = {
